% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pebyg}
\alias{pebyg}
\title{estimate the pe by growth}
\usage{
pebyg(growth, r = 0.1, fg = 0.05)
}
\arguments{
\item{growth}{, the growth rate in the following years,
eg c(0.5,0.3,0.3) means 0.5, 0.30 ,0.30 growth in the following
3 years}

\item{r}{return rate , 0.10 by default}

\item{fg}{forever growth , 0.05 by default}
}
\description{
estimate the pe by growth
}
\details{
estimate the pe by growth
}
\author{
xuyang
}
