% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pebyg2}
\alias{pebyg2}
\title{pebyg2}
\usage{
pebyg2(growth, ...)
}
\arguments{
\item{growth}{}

\item{...}{}
}
\description{
calc pe according to the growth
}
\details{
the growth is based on the latest year,  c(0.3,0.6,0.9)
will mean the real income is (1.3,1.6,1.9) in the following
3 years.
}
\author{
xuyang
}
